refactor: Normalize module path
