+++ lib/plugins/aws/rollback.js
@@ -1,11 +1,11 @@
 'use strict';
 
 const BbPromise = require('bluebird');
-const validate = require('../lib/validate');
-const setBucketName = require('../lib/setBucketName');
-const updateStack = require('../lib/updateStack');
-const monitorStack = require('../lib/monitorStack');
-const findAndGroupDeployments = require('../utils/findAndGroupDeployments');
+const validate = require('./lib/validate');
+const setBucketName = require('./lib/setBucketName');
+const updateStack = require('./lib/updateStack');
+const monitorStack = require('./lib/monitorStack');
+const findAndGroupDeployments = require('./utils/findAndGroupDeployments');
 
 class AwsRollback {
   constructor(serverless, options) {
