+++ lib/plugins/index.js
@@ -30,7 +30,7 @@ module.exports = [
   require('./aws/logs.js'),
   require('./aws/metrics.js'),
   require('./aws/remove/index.js'),
-  require('./aws/rollback/index.js'),
+  require('./aws/rollback.js'),
   require('./aws/rollbackFunction/index.js'),
   require('./aws/package/compile/layers.js'),
   require('./aws/package/compile/functions.js'),
